move api test into separate package
