+++ src/test/scala/com/campudus/tableaux/api/PerformanceTest.scala
@@ -1,11 +1,12 @@
-package com.campudus.tableaux
+package com.campudus.tableaux.api
 
 import com.campudus.tableaux.database.model.TableauxModel._
+import com.campudus.tableaux.testtools.TableauxTestBase
 import io.vertx.core.json.JsonObject
 import io.vertx.ext.unit.TestContext
 import io.vertx.ext.unit.junit.VertxUnitRunner
-import org.junit.{Ignore, Test}
 import org.junit.runner.RunWith
+import org.junit.{Ignore, Test}
 import org.vertx.scala.core.json.Json
 
 import scala.collection.JavaConverters._
