+++ src/test/scala/com/campudus/tableaux/api/content/MultiLanguageTest.scala
@@ -1,6 +1,7 @@
-package com.campudus.tableaux
+package com.campudus.tableaux.api.content
 
 import com.campudus.tableaux.database.model.TableauxModel.{ColumnId, RowId, TableId}
+import com.campudus.tableaux.testtools.{TableauxTestBase, TestCustomException}
 import io.vertx.ext.unit.TestContext
 import io.vertx.ext.unit.junit.VertxUnitRunner
 import org.junit.Test
