+++ src/test/scala/com/campudus/tableaux/api/structure/HiddenTableTest.scala
@@ -1,5 +1,6 @@
-package com.campudus.tableaux
+package com.campudus.tableaux.api.structure
 
+import com.campudus.tableaux.testtools.TableauxTestBase
 import io.vertx.ext.unit.TestContext
 import io.vertx.ext.unit.junit.VertxUnitRunner
 import org.junit.Test
