+++ src/test/scala/com/campudus/tableaux/controller/StructureControllerTest.scala
@@ -1,9 +1,9 @@
 package com.campudus.tableaux.controller
 
-import com.campudus.tableaux.TableauxTestBase
 import com.campudus.tableaux.database.domain.CreateSimpleColumn
 import com.campudus.tableaux.database.model.StructureModel
 import com.campudus.tableaux.database.{DatabaseConnection, SingleLanguage, TextType}
+import com.campudus.tableaux.testtools.TableauxTestBase
 import io.vertx.ext.unit.TestContext
 import io.vertx.ext.unit.junit.VertxUnitRunner
 import io.vertx.scala.SQLConnection
