+++ src/test/scala/com/campudus/tableaux/controller/SystemControllerTest.scala
@@ -1,7 +1,7 @@
 package com.campudus.tableaux.controller
 
-import com.campudus.tableaux.TableauxTestBase
 import com.campudus.tableaux.router.SystemRouter
+import com.campudus.tableaux.testtools.TableauxTestBase
 import io.vertx.ext.unit.TestContext
 import io.vertx.ext.unit.junit.VertxUnitRunner
 import org.junit.Test
