+++ src/test/scala/com/campudus/tableaux/database/model/FileModelTest.scala
@@ -1,9 +1,9 @@
 package com.campudus.tableaux.database.model
 
-import com.campudus.tableaux.TableauxTestBase
 import com.campudus.tableaux.controller.MediaController
 import com.campudus.tableaux.database.DatabaseConnection
 import com.campudus.tableaux.database.domain.MultiLanguageValue
+import com.campudus.tableaux.testtools.TableauxTestBase
 import io.vertx.ext.unit.TestContext
 import io.vertx.ext.unit.junit.VertxUnitRunner
 import io.vertx.scala.SQLConnection
